[refactor] 채팅 관련 기능 디렉토리 이동
