+++ src/main/java/potatoes/server/chat/controller/ChatController.java
@@ -1,4 +1,4 @@
-package potatoes.server.controller;
+package potatoes.server.chat.controller;
 
 import static org.springframework.http.MediaType.*;
 
@@ -20,6 +20,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import potatoes.server.chat.service.ChatService;
 import potatoes.server.constant.ChatSortType;
 import potatoes.server.dto.ChatOverviewResponse;
 import potatoes.server.dto.ChatSummaryResponse;
@@ -27,7 +28,6 @@ import potatoes.server.dto.CommonResponse;
 import potatoes.server.dto.MarkAsReadPublish;
 import potatoes.server.dto.MessagePublish;
 import potatoes.server.dto.RecentChatResponse;
-import potatoes.server.service.ChatService;
 import potatoes.server.utils.annotation.Authorization;
 import potatoes.server.utils.stomp.StompUserPrincipal;
