+++ src/main/java/potatoes/server/chat/entity/Chat.java
@@ -1,4 +1,4 @@
-package potatoes.server.entity;
+package potatoes.server.chat.entity;
 
 import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
@@ -14,6 +14,8 @@ import jakarta.persistence.OneToOne;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import potatoes.server.entity.Travel;
+import potatoes.server.entity.User;
 
 @Getter
 @NoArgsConstructor(access = PROTECTED)
