+++ src/main/java/potatoes/server/chat/entity/ChatMessageUser.java
@@ -1,4 +1,4 @@
-package potatoes.server.entity;
+package potatoes.server.chat.entity;
 
 import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
@@ -12,11 +12,12 @@ import jakarta.persistence.ManyToOne;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import potatoes.server.entity.User;
 
 @Getter
 @NoArgsConstructor(access = PROTECTED)
 @Entity
-public class ChatUser {
+public class ChatMessageUser {
 
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
@@ -27,12 +28,23 @@ public class ChatUser {
 	private Chat chat;
 
 	@ManyToOne(fetch = LAZY)
+	@JoinColumn(name = "chat_message_id")
+	private ChatMessage chatMessage;
+
+	@ManyToOne(fetch = LAZY)
 	@JoinColumn(name = "user_id")
 	private User user;
 
+	private boolean hasRead;
+
 	@Builder
-	public ChatUser(Chat chat, User user) {
+	public ChatMessageUser(Chat chat, ChatMessage chatMessage, User user) {
 		this.chat = chat;
+		this.chatMessage = chatMessage;
 		this.user = user;
+	}
+
+	public void markAsRead() {
+		hasRead = true;
 	}
 }
