+++ src/main/java/potatoes/server/chat/repository/ChatImageRepository.java
@@ -1,4 +1,4 @@
-package potatoes.server.repository;
+package potatoes.server.chat.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import potatoes.server.entity.ChatImage;
+import potatoes.server.chat.entity.ChatImage;
 
 public interface ChatImageRepository extends JpaRepository<ChatImage, Long> {
