+++ src/main/java/potatoes/server/chat/repository/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package potatoes.server.repository;
+package potatoes.server.chat.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import potatoes.server.entity.ChatMessage;
+import potatoes.server.chat.entity.ChatMessage;
 
 public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
