+++ src/main/java/potatoes/server/chat/service/ChatService.java
@@ -1,4 +1,4 @@
-package potatoes.server.service;
+package potatoes.server.chat.service;
 
 import static java.util.Comparator.*;
 import static potatoes.server.constant.AlarmStatus.*;
@@ -21,6 +21,16 @@ import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import potatoes.server.chat.entity.Chat;
+import potatoes.server.chat.entity.ChatImage;
+import potatoes.server.chat.entity.ChatMessage;
+import potatoes.server.chat.entity.ChatMessageUser;
+import potatoes.server.chat.entity.ChatUser;
+import potatoes.server.chat.repository.ChatImageRepository;
+import potatoes.server.chat.repository.ChatMessageRepository;
+import potatoes.server.chat.repository.ChatMessageUserRepository;
+import potatoes.server.chat.repository.ChatRepository;
+import potatoes.server.chat.repository.ChatUserRepository;
 import potatoes.server.constant.ChatSortType;
 import potatoes.server.dto.AlarmSubscribe;
 import potatoes.server.dto.ChatAlbumResponse;
@@ -31,19 +41,9 @@ import potatoes.server.dto.MessagePublish;
 import potatoes.server.dto.MessageSubscribe;
 import potatoes.server.dto.ParticipantsInfoResponse;
 import potatoes.server.dto.RecentChatResponse;
-import potatoes.server.entity.Chat;
-import potatoes.server.entity.ChatImage;
-import potatoes.server.entity.ChatMessage;
-import potatoes.server.entity.ChatMessageUser;
-import potatoes.server.entity.ChatUser;
 import potatoes.server.entity.TravelUser;
 import potatoes.server.entity.User;
 import potatoes.server.error.exception.WeGoException;
-import potatoes.server.repository.ChatImageRepository;
-import potatoes.server.repository.ChatMessageRepository;
-import potatoes.server.repository.ChatMessageUserRepository;
-import potatoes.server.repository.ChatRepository;
-import potatoes.server.repository.ChatUserRepository;
 import potatoes.server.repository.TravelUserRepository;
 import potatoes.server.utils.s3.S3UtilsProvider;
 import potatoes.server.utils.stomp.StompUserPrincipal;
