+++ src/main/java/potatoes/server/config/SwaggerConfig.java
@@ -15,8 +15,8 @@ import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.security.SecurityRequirement;
 import io.swagger.v3.oas.models.security.SecurityScheme;
+import potatoes.server.chat.controller.ChatController;
 import potatoes.server.controller.AuthController;
-import potatoes.server.controller.ChatController;
 import potatoes.server.controller.MailController;
 import potatoes.server.controller.ReviewController;
 import potatoes.server.controller.TravelController;
