+++ src/main/java/potatoes/server/service/TravelService.java
@@ -18,6 +18,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import potatoes.server.chat.entity.Chat;
+import potatoes.server.chat.repository.ChatRepository;
 import potatoes.server.constant.ParticipantRole;
 import potatoes.server.constant.TravelSortType;
 import potatoes.server.constant.TravelStatus;
@@ -31,14 +33,12 @@ import potatoes.server.dto.TravelListResponse;
 import potatoes.server.dto.TravelPlanResponse;
 import potatoes.server.dto.TravelSummaryResponse;
 import potatoes.server.entity.Bookmark;
-import potatoes.server.entity.Chat;
 import potatoes.server.entity.Travel;
 import potatoes.server.entity.TravelPlan;
 import potatoes.server.entity.TravelUser;
 import potatoes.server.entity.User;
 import potatoes.server.error.exception.WeGoException;
 import potatoes.server.repository.BookmarkRepository;
-import potatoes.server.repository.ChatRepository;
 import potatoes.server.repository.TravelPlanRepository;
 import potatoes.server.repository.TravelRepository;
 import potatoes.server.repository.TravelUserRepository;
