+++ src/main/java/potatoes/server/utils/stomp/ChatEventHandler.java
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import potatoes.server.entity.ChatUser;
+import potatoes.server.chat.entity.ChatUser;
+import potatoes.server.chat.repository.ChatRepository;
+import potatoes.server.chat.repository.ChatUserRepository;
 import potatoes.server.entity.User;
 import potatoes.server.error.exception.WeGoException;
-import potatoes.server.repository.ChatRepository;
-import potatoes.server.repository.ChatUserRepository;
 import potatoes.server.utils.jwt.JwtTokenUtil;
 
 @Slf4j
